extracted/wa/waldur-api-client/pyproject.toml
[tool.poetry]
name = "waldur-api-client"
version = "8.0.6"
description = "A client library for accessing Waldur API"
authors = ["OpenNode Team <info@opennodecloud.com>"]
license = "MIT"
homepage = "https://waldur.com"
documentation = "https://docs.waldur.com"
readme = "README.md"
packages = [
    {include = "waldur_api_client"},
]
include = [
    "CHANGELOG.md",
    "waldur_api_client/py.typed",
    "waldur_api_client/llms.txt",
    "waldur_api_client/llms-full.txt",
    "waldur_api_client/api-map.md",
]

[tool.poetry.dependencies]
python = "^3.9"
httpx = ">=0.20.0,<0.29.0"
attrs = ">=22.2.0"
python-dateutil = "^2.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["F", "I", "UP"]
